remove unused fs calls
