+++ scraper/consts.js
@@ -1,14 +1,6 @@
-export const DEFAULT_INFILE = "./data/games.json";
-export const DEFAULT_OUTFILE = "./data/games.json";
-export const APPLIST_FILE = "./data/applist.json";
-export const DISCARTED_FILE = "./data/discarted.json";
-export const NOTRELEASED_FILE = "./data/notreleased.json";
 export const DEFAULT_SLEEP = 1.5;
 export const DEFAULT_RETRIES = 4;
-export const DEFAULT_AUTOSAVE = 100;
 export const DEFAULT_TIMEOUT = 10000;
-export const DEFAULT_CURRENCY = "us";
-export const DEFAULT_LANGUAGE = "en";
 export const LOG_ICON = ["i", "W", "E", "!"];
 export const INFO = 0;
 export const WARNING = 1;
