+++ scraper/utils.js
@@ -79,6 +79,77 @@ export const getGamePkgs = (app) => {
   return packages;
 };
 
+const getDate = (app) => {
+  try {
+    return app.release_date && !app.release_date.coming_soon
+      ? new Date(app.release_date.date).toISOString()
+      : "";
+  } catch {
+    return "";
+  }
+};
+
+export function shuffle(array) {
+  let currentIndex = array.length;
+
+  // While there remain elements to shuffle...
+  while (currentIndex != 0) {
+    // Pick a remaining element...
+    let randomIndex = Math.floor(Math.random() * currentIndex);
+    currentIndex--;
+
+    // And swap it with the current element.
+    [array[currentIndex], array[randomIndex]] = [
+      array[randomIndex],
+      array[currentIndex],
+    ];
+  }
+}
+
+export function parseSteamGame(app) {
+  return {
+    name: app.name.trim(),
+    release_date: getDate(app),
+    dlc_count: app.dlc ? app.dlc.length : 0,
+    detailed_description: app.detailed_description?.trim() ?? "",
+    price:
+      app.is_free || !app.price_overview
+        ? 0.0
+        : priceToFloat(app.price_overview.final_formatted),
+    about_the_game: app.about_the_game?.trim() ?? "",
+    short_description: app.short_description?.trim() ?? "",
+    reviews: app.reviews?.trim() ?? "",
+    platforms: {
+      windows: app.platforms?.windows,
+      mac: app.platforms?.mac,
+      linux: app.platforms?.linux,
+    },
+    header_image: app.header_image?.trim() ?? "",
+    website: app.website?.trim() ?? "",
+    metacritic_score: app.metacritic ? parseInt(app.metacritic.score) : 0,
+    metacritic_url: app.metacritic?.url ?? "",
+    achievements: app.achievements ? parseInt(app.achievements.total) : 0,
+    recommendations: app.recommendations?.total ?? 0,
+    notes: app.content_descriptors?.notes ?? "",
+    developers: app.developers?.map((developer) => developer.trim()),
+    publishers: app.publishers?.map((publisher) => publisher.trim()),
+    categories: app.categories?.map((category) =>
+      category?.description?.trim()
+    ),
+    genres: app.genres?.map((genres) => genres?.description?.trim()),
+    screenshots: app.screenshots?.map((screenshot) => screenshot?.path_full),
+    detailed_description: sanitizeText(app.detailed_description),
+    about_the_game: sanitizeText(app.about_the_game),
+    short_description: sanitizeText(app.short_description),
+    reviews: sanitizeText(app.reviews),
+    notes: sanitizeText(app.notes),
+    movies: app.movies?.map((movie) => movie?.mp4?.max),
+    ...getLanguages(app),
+    packages: getGamePkgs(app),
+    adult_game: app.ratings?.steam_germany?.banned === "1",
+  };
+}
+
 export async function doRequest(
   url,
   parameters = null,
